リファクタリング: HTMLをtemplates.pyに分離、コードモジュール化
